# swagger/pets/openapi.yaml
# [START swagger]
swagger: "2.0"
info:
  description: "A api example"
  title: "Endpoints Example"
  version: "1.0.0"
host: "localhost:3555"
# [END swagger]
basePath: "/api/"
consumes:
- "application/json"
produces:
- "application/json"
schemes:
- "http"
paths:
  "/echo":
    post:
      description: "Echo back a given message."
      operationId: "echo"
      produces:
      - "application/json"
      responses:
        200:
          description: "Echo"
          schema:
            $ref: "#/definitions/echoMessage"
      parameters:
      - description: "Message to echo"
        in: body
        name: message
        required: true
        schema:
          $ref: "#/definitions/echoMessage"
  "/pets":
    get:
      description: "Gets all pets"
      operationId: "pets"
      produces:
      - "application/json"
      responses:
        200:
          description: "A list of pets"
          schema:
            type: array
            items:
              $ref: "#/definitions/pet"
    post:
      description: "Create a pet"
      operationId: "pets"
      produces:
      - "application/json"
      responses:
        200:
          description: "Returns the id of the newly created pet"
          schema:
            type: string
      parameters:
      - description: "The pet to insert"
        in: body
        name: InsertPet
        required: true
        schema:
          $ref: "#/definitions/pet"
  "/pets/{id}":
    get:
      description: "Returns a pet based by id"
      operationId: "find pet by id"
      produces:
      - "application/json"
      parameters:
      - name: id
        in: path
        description: Id of pet to fetch
        required: true
        type: string
      responses:
        200:
          description: "pet response"
          schema:
              $ref: "#/definitions/pet"
definitions:
  echoMessage:
    properties:
      message:
        type: "string"
  pet:
    properties:
      id:
        type: "string"
      type:
        type: "string"
      name:
        type: "string"